read_IPMI_info.cc reads through each slot on shelf and prints useful board/product information

+++ include/ApolloReader.hh
@@ -53,14 +53,14 @@ private:
   
   std::vector<uint8_t> header;
   // fields for data stored in header
-  uint8_t headerFormatVersion;
-  uint8_t internalUseStartingOffset;
-  uint8_t chassisInfoStartingOffset;
-  uint8_t boardStartingOffset;
-  uint8_t productInfoStartingOffset;
-  uint8_t multiRecordStartingOffset;
-  uint8_t pad;
-  uint8_t commonHeaderChecksum;
+  int headerFormatVersion;
+  int internalUseStartingOffset;
+  int chassisInfoStartingOffset;
+  int boardStartingOffset;
+  int productInfoStartingOffset;
+  int multiRecordStartingOffset;
+  int pad;
+  int commonHeaderChecksum;
 
   // raw hex data from the blade
   std::vector<uint8_t> data;
